Updated lecture notes for Lecture 3
The readme.md for the third lecture (datatypes and variables) was out of date. I updated it to reflect my most recent version of the lecture, corresponding to my slides (which I also added). Also, note that our current pandoc setup can't handle \frac commands in embedded math ($...$ regions), so I just had to write division operators for the fractions in this text.