Minor textual edits, added link to != TT
I briefly mention the concept of XOR because wikipedia doesn't have a truth table for inequality but the one for XOR holds in this case.  I put a warning that conceptually XOR is different despite the common truth table in this case.  I didn't elaborate on what XOR is and how it differs from inequality because I didn't want to confuse students (as those differences are minimized by the 2-value nature of boolean logic).