Potential fix for issue #50
I moved the text on foreach, break, and continue from the end of Lecture 15 to the beginning of Lecture 16, so that there's only one section that contains this information. Alternatively we could delete Lecture 16 entirely and put it all back in Lecture 15, but this seems like the easiest fix for now since it doesn't require renumbering anything.