Promoted some bullet points to subheadings
Many of my first-level bullet points were written more like section
headers, and it would be easier for students to navigate if they were
actually formatted as headings in the Markdown. For now I'm going to try
making them level-4 headings, even though this skips level 3, because I
don't want the text to be too big.