Fixed spacing in truth table code
These code snippets seem to have been written with the assumption that \t produced exactly 4 spaces, but in my tests (at least in Windows) it actually means "move up to the next tab stop" and the amount of space inserted varies depending on the number of characters on the line so far. This version produces correctly-aligned tables for me, by only using \t after variable-length words (not before).